+++ todo-or-die/example/simple_usage_example.dart
@@ -8,7 +8,11 @@ void main() {
   );
 
   // with conditions
-  todoOrDie('Rewrite this to scale better', given: () => userCount() > 1000);
+  todoOrDie(
+    'Rewrite this to scale better',
+    given: () => userCount() >= 1000,
+    givenDesription: 'we have reached 1000 users',
+  );
 }
 
 int userCount() => 0; // function to get user count